+++ octasine/src/audio/gen/lfo.rs
@@ -1,6 +1,6 @@
 use arrayvec::ArrayVec;
 
-use crate::audio::parameters::*;
+use crate::audio::parameters::{common::AudioParameter, AudioParameterLfo};
 use crate::audio::voices::lfos::VoiceLfo;
 use crate::common::*;
 use crate::parameter_values::lfo_target::*;